templates/audius-base/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.4",
  "uvicorn[standard]==0.32.0",
  "sqlalchemy[asyncio]==2.0.36",
  "asyncpg==0.30.0",
  "alembic==1.14.0",
  "pydantic==2.9.2",
  "pydantic-settings==2.6.1",
  "python-jose[cryptography]==3.3.0",
  "httpx==0.27.2",
]

[tool.ruff]
line-length = 88
target-version = "py311"
